requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
rich>=13.7.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
kafka-python>=2.0.2
pydantic>=2.8.0
mcp>=1.0.0
httpx>=0.27.0
httpx-sse>=0.4.0
anyio>=4.4.0
